refactor: unify HTTP response handling into HttpResponse for RESTful consistency
