+++ src/Controllers/UserController.php
@@ -4,29 +4,29 @@ namespace App\Controllers;
 
 use App\Http\Request;
 use App\Http\Response;
+use App\Http\HttpResponse;
 use App\Services\UserService;
 
 class UserController
 {
+    private $httpResponse;
+
+    public function __construct(Request $request, Response $response)
+    {
+        $this->httpResponse = new HttpResponse($request, $response);
+    }
+
     public function store(Request $request, Response $response)
     {
         $body = Request::body();
 
         $userService = UserService::create($body);
 
-        if (isset($userService['error'])) {
-            return $response->json([
-                'error' => true,
-                'success' => false,
-                'message' => $userService['error']
-            ], 400);
+        if (isset($userService['error']) && $userService['error'] === true) {
+            return $this->httpResponse->badRequest($userService['message']);
         }
 
-        $response->json([
-            'error' => false,
-            'success' => true,
-            'data' => $userService
-        ], 201);
+        $this->httpResponse->created($userService);
     }
 
     public function login(Request $request, Response $response)
